feat(event-display): restructure phoenix-event-display
