+++ packages/phoenix-event-display/src/managers/three-manager/animations-manager.ts
