+++ packages/phoenix-event-display/src/managers/three-manager/controls-manager.ts
