+++ packages/phoenix-event-display/src/managers/three-manager/effects-manager.ts
