+++ packages/phoenix-event-display/src/managers/three-manager/fonts/helvetiker_regular.typeface.json
