+++ packages/phoenix-event-display/src/managers/three-manager/renderer-manager.ts
