+++ packages/phoenix-event-display/src/managers/ui-manager/phoenix-menu/phoenix-menu-node.ts
