+++ packages/phoenix-event-display/src/tests/info-logger.spec.ts
@@ -1,4 +1,4 @@
-import { InfoLogger } from '../info-logger';
+import { InfoLogger } from '../helpers/info-logger';
 
 describe('InfoLogger', () => {
   let infoLogger: InfoLogger;
